Transaction History, My portfolio, News Articles pages finished, UI improvements

+++ publicServer/DataCollector/Commands/CommandList/GetLatestNews.py
@@ -4,33 +4,68 @@ import requests
 from sqlalchemy.exc import OperationalError
 
 from publicServer.DataCollector.Commands.Command import Command
+from publicServer.DataCollector.Database.Models.Company import Company
 from publicServer.DataCollector.Database.Models.LatestNew import LatestNew
 from publicServer.DataCollector.Database.session import db
-from publicServer.config.constants import API_ENDPOINT, TEN_MINUTES
+from publicServer.config.constants import API_ENDPOINT, ONE_HOUR
 from publicServer.config.definitions import KEY_URL
 
 
 class GetLatestNews(Command, ABC):
     def __init__(self):
-        super().__init__(TEN_MINUTES, 5, 1)
-        self.url = API_ENDPOINT + "v3/stock_news?limit=10&" + KEY_URL
+        super().__init__(ONE_HOUR, 5, 1)
+        self.url = API_ENDPOINT + "v3/stock_news?tickers={}&page={}&" + KEY_URL
+        print(self.url)
 
     def execute(self):
+        self.prepareRequest("0")
         result = requests.get(self.url)
         if result.status_code == 200:
-            self.collectData(result.json())
+            self.collectData(result.json(), True)
         else:
             print(f"[GetLatestNews] failed to get latest news {result.status_code}")
 
+        self.prepareRequest("1")
+        result = requests.get(self.url)
+        if result.status_code == 200:
+            self.collectData(result.json(), False)
+        else:
+            print(f"[GetLatestNews] failed to get latest news {result.status_code}")
+
+        self.prepareRequest("2")
+        result = requests.get(self.url)
+        if result.status_code == 200:
+            self.collectData(result.json(), False)
+        else:
+            print(f"[GetLatestNews] failed to get latest news {result.status_code}")
+
+        self.prepareRequest("3")
+        result = requests.get(self.url)
+        if result.status_code == 200:
+            self.collectData(result.json(), False)
+        else:
+            print(f"[GetLatestNews] failed to get latest news {result.status_code}")
+
+    def prepareRequest(self, page: str):
+        self.url = API_ENDPOINT + "v3/stock_news?tickers={}&page={}&" + KEY_URL
+        ticker_list = ""
+        for ticker in db.query(Company.ticker).all():
+            ticker_list += ticker[0] + ","
+        self.url = self.url.format(ticker_list[:-1], page)
+
     @staticmethod
-    def collectData(result):
+    def collectData(result, delete: bool):
         try:
-            db.query(LatestNew).delete()
+            if delete:
+                db.query(LatestNew).delete()
+                db.commit()
         except OperationalError as exc:
             print(exc.args)
-        print(db.query(LatestNew).all())
         for new in result:
             new_to_insert = LatestNew(ticker=new["symbol"], title=new["title"], image=new["image"],
-                                      site=new["site"], text=new["text"], url=new["url"])
-            db.add(new_to_insert)
-        db.commit()
+                                      site=new["site"], text=new["text"], url=new["url"], date=new["publishedDate"])
+            try:
+                db.add(new_to_insert)
+                db.commit()
+            except Exception as e:
+                db.rollback()
